refactor: reapply solid principles to placeShip
refactored placeShip mechanic to no longer pass the ship object itself
to the board at index, but instead pass a shipIndex reference as a
number (previously was passing the entire object) while also refactoring
the way the receiveAttack method therefore to access the number passed
instead of the object; and refactored the getShip helper method to also
target the index in line with abovementioned refactoring.

refactored createShip modlue to return the hits starting from 0 and
incrementing every time the ship gets hit. this previously decremented
down from the ships's length down to zero. and last but not least,
refactored isSunk logic to work in line with this change to evaluate to
true once the hits reaches >= to the ship's length value, instead of
being decremented down to zero.

also refactored the unit tests that were targetting these methods to
make sure they are expecting the correct behaviour in light of these
changes.