fix: 2 methods logic & refactor several methods
- the recveiveAttack method logic was fixed and verified to pass the
tests that it now correctly calls the ship.hit()
- the isAllShipsSunken method -this method will be renamed to a more
appropriate name in the next commit- was refactored to fix the falsy
logic to now returning true only if each ship in the array returns
isSunk true. (it was unreliably returning seemingly random bool before)

- furthermore, all of the smaller helper methods that were not defined
in the functions body are now moved up and are defined properly to be
used internally inside the function to keep things dry.
- other methods, where the herlper methods were used internally were
also refactored to accomodate the changes and the general theme.

necessary tests to verify these changes were written and passed.