feat: add base ui & refactor: logic and display
the `GameController` and `DomController` have been refactored to utilize
a third module `AppController` as a middleman between the two.

the basic ui was implemented and is now working with click handlers
performing attacks correctly and updating the cells on the dom.

several additional refactors were performed where deemed necessary to
modularize the app and strive for more solid-ness.