feat: add `CpuPlayer` logic & DOM updates
the cpu player can now perform attacks on random rows and cols. the game
alternates between player1 and player2, temporarily disabling the board
of the opposing player based on the `currentTurn`. the cpu player doesnt
know if a cell was already attack before, so it doesnt determine that
for the time being.

fix: `Gameboard.receiveAttack()` no longer throws an error, and instead,
gracefully logs a message to the console (for the time being) & returns
the error message as string. this was done to prevent annoying errors
being thrown left and right every time the cpu player attempts to attack
a cell that was already attacked before. perhaps in the future, i could
go back to actually throwing an error if deemed necessary.

also updated the relevant unit test to accomodate for this temporary
change.