feat: add cpuPlayer & refactor: `GameController`
the computer player functionality was improved upon. now the cpu does
indeed play what it is it's turn. the `runGame` function was refactored
greatly to fix an issue where the underlying logic dictated the attack
was always performed on the `currentPlayer`'s `ownBoard` instead of the
`opponent`'s board. several other helper methods and DOM methods were
refactored to align with these changes.