feat: improve the logic of `computerPlayer`
the `computerPlayer` now tries to attack the adjacent cells if it lands
a hit on one of the ships of the opposing player. the core logic seems
to be working "ok" as of right now but it fails to actually attack to
the last cell of the ship, therefore never managing to sink it.

further logic to fix this major issue is due in the next commit.