feat: add `Set` to store attacked cells
the `ComputerPlayer` factory now has two new methods to store attacked
cells in. the next random coordiante genetation will first delegate to
the Set and try to eliminate the included coords. However, this logic
still bugs out once the number of available cells get under a threshold.

i have not yet been able to identify the cause of the faulty behaviour,
i will look into it more, come next commit, while also trying to figure
a way to implement direction-awareness to the `ComputerPlayer`logic so
that once it discovers the ship's orientation, it will only target the
adjacent cells in that orientation.