fix: issue with detect axis, `lastHit` & targets
the way the `ComputerPlayer` detects ship orientation, store `lastHit`,
and add `addTargetsAlongOrientation` were completely broken, resulting
in the gameflow getting stuck because of this. this issue is now largely
fixed, where, even though the `ComputerPlayer` is not 100% optimized to
find the correct cells to attack to, but now, it rarely gets stuck and
causes the game flow to be stuck.

it now detects the orientation correctly, and pursues in that direction
until either it -coincidentally- sinks the ship, or it reaches the edge
of the gameBoard. once either one of these happens, the ComputerPlayer
then resets its mode back to hunting mode, and continues with random
attacks until it hits another ship.

the issue with the logic of getting adjacent cells in the detected axis,
where the ComputerPlayer is not backtracking once it hits the last cell
of the ship in the current direction is currently known and a fix is due
in the next commit.