fix: add backtracking to the ai player
the `ComputerPlayer` now knows how to backtrack and does so proper. it
no longer goes all the way to the edge of the game field before turning
back to finish up the remaining cells of the current ship. this marks
the completion the the ai player implementation for now. next commit
will possibly see the implementation of ship placement for both players.