feat: add manuel ship placement logic and ui
the user can now place their ships one by one (for a total of 5 ships
in deterministic order) to their own board. for the time being, this
only supports horizontal placements, but a switch between either axis
will be implemented in the next commit.

it will be followed by the implementation for the `computerPlayer` to
place ships randomly soon after.